refactor some `FnType` stuff to `rustc::ty::layout`
